修复警告

+++ Classes/FanCalculator/FanCalculatorScene.cpp
@@ -146,7 +146,7 @@ void FanCalculatorScene::calculate() {
 
     std::sort(param.hand_tiles.standing_tiles, param.hand_tiles.standing_tiles + param.hand_tiles.tile_count);
 
-    param.flower_count = static_cast<uint16_t>(flowerCnt);
+    param.flower_count = static_cast<uint8_t>(flowerCnt);
     mahjong::fan_table_t fan_table = { 0 };
 
     // 获取绝张、杠开、抢杠、海底信息
